requirements.txt
# Spotify API & HTTP requests
requests==2.32.3

# Environment variable management
python-dotenv==1.0.1

# Data manipulation
pandas==2.2.2
numpy==1.26.4

# Data visualization
matplotlib==3.9.0
seaborn==0.13.2

# Machine learning (for final project phase)
scikit-learn==1.5.0

# Jupyter notebook
jupyter==1.0.0
ipykernel==6.29.4

# Web scraping (for Billboard / WhoSampled)
beautifulsoup4==4.12.3
lxml==5.2.2

# Kaggle dataset download
kaggle==1.6.14